requirements.txt
websockets==7.0
paho_mqtt==1.4.0
